Created first opengl display window
Added a library for  glad to load a modern OpenGL 3.3 API
Created a shader helper outline based on code found online that can
load shaders and link them to a program for use later.
Added two basic shader files that are empty right now to be filled in later.
Initialized glfw and created the first window loop to display a window
with a blue hue.
Added event handlers to capture mouse movement and key presses and handled
esc press to close out the window.